refactor: detangle list
flyer list should not hijack normal list item
