+++ packages/styleguide/src/components/Typography/fontRules.js
@@ -17,7 +17,3 @@ export const interactionFontRule = css({
     textDecoration: `underline wavy ${colors.error}`,
   },
 })
-
-export const flyerFontRule = css({
-  ...fontStyles.sansSerifMedium,
-})
